requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
click==8.2.1
colorama==0.4.6
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
numpy==2.2.6
pandas==2.3.1
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
